create file_schemas directory to move pydantic schemas to and update fs_pred accordingly
